+++ examples/complete/main.tf
@@ -92,13 +92,21 @@ module "s3_bucket" {
     target_prefix = "log/"
   }
 
-  cors_rule = {
-    allowed_methods = ["PUT", "POST"]
-    allowed_origins = ["https://modules.tf", "https://terraform-aws-modules.modules.tf"]
-    allowed_headers = ["*"]
-    expose_headers  = ["ETag"]
-    max_age_seconds = 3000
-  }
+  cors_rule = [
+    {
+      allowed_methods = ["PUT", "POST"]
+      allowed_origins = ["https://modules.tf", "https://terraform-aws-modules.modules.tf"]
+      allowed_headers = ["*"]
+      expose_headers  = ["ETag"]
+      max_age_seconds = 3000
+      }, {
+      allowed_methods = ["PUT"]
+      allowed_origins = ["https://example.com"]
+      allowed_headers = ["*"]
+      expose_headers  = ["ETag"]
+      max_age_seconds = 3000
+    }
+  ]
 
   lifecycle_rule = [
     {
@@ -174,4 +182,10 @@ module "s3_bucket" {
       }
     }
   }
+
+  // S3 bucket-level Public Access Block configuration
+  block_public_acls       = true
+  block_public_policy     = true
+  ignore_public_acls      = true
+  restrict_public_buckets = true
 }