+++ versions.tf
@@ -1,0 +1,7 @@
+terraform {
+  required_version = ">= 0.12.6, < 0.14"
+
+  required_providers {
+    aws = ">= 3.0, < 4.0"
+  }
+}